Add files via upload
The logic to make a line follower is to detect the black line by making use of two IR sensors at the front on the left and right sides of the car. These sensors give analog as well as digital signals, we are making use of ADC values to differentiate between the black and white line. It shows the ADC value 1023 for a black line and a very low reading (50-100) for any other colors, hence whenever a black line crosses the sensor gives a 1023 ADC value intrusting that line is black, and accordingly the motor speed and directions are set to take turns or move forward.